backend/api/pyproject.toml
[project]
name = "kineo-coach-api"
version = "0.1.0"
description = "Backend API para Kineo Coach"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "email-validator>=2.3.0,<3.0.0",
  "fastapi>=0.116.0,<1.0.0",
  "openai>=1.109.0,<2.0.0",
  "PyJWT>=2.10.1,<3.0.0",
  "psycopg[binary]>=3.2.9,<4.0.0",
  "pydantic-settings>=2.10.0,<3.0.0",
  "python-dotenv>=1.1.0,<2.0.0",
  "sqlalchemy>=2.0.41,<3.0.0",
  "uvicorn[standard]>=0.35.0,<1.0.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.28.0,<1.0.0",
  "pytest>=8.4.0,<9.0.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
